feat: #33 Add smart contract for dispersing $ELIMU tokens
Skip ABI generation when dry-run

+++ disperse/migrations/1_deploy_contract.js
@@ -4,12 +4,15 @@ const fs = require('fs');
 module.exports = function (deployer, network) {
   console.log('network:', network)
 
+  let isDryRun = network.endsWith('-fork')
+  if (!isDryRun) {
+    // Save the contract's ABI
+    const contractJson = JSON.parse(fs.readFileSync('./build/contracts/' + ElimuDisperse.contractName + '.json', 'utf8'));
+    console.log('Saving ABI:', './abis/' + network + '/' + ElimuDisperse.contractName + '.json.abi')
+    fs.writeFileSync('./abis/' + network + '/' + ElimuDisperse.contractName + '.json.abi', JSON.stringify(contractJson.abi))
+  }
+
   // Deploy the contract
   console.log('contractName:', ElimuDisperse.contractName)
   deployer.deploy(ElimuDisperse);
-
-  // Save the contract's ABI
-  const contractJson = JSON.parse(fs.readFileSync('./build/contracts/' + ElimuDisperse.contractName + '.json', 'utf8'));
-  console.log('Saving ABI:', './abis/' + network + '/' + ElimuDisperse.contractName + '.json.abi')
-  fs.writeFileSync('./abis/' + network + '/' + ElimuDisperse.contractName + '.json.abi', JSON.stringify(contractJson.abi))
 };
